Replace FligthFilter interface: use Predicate<Flight>
So we can pass it directly to Stream.filter

Created the package es.danielcr86.flightSearch.flightPredicates.
All the logic intended to filter flights should be gathered here.
The idea is to have static functions that return an instance of
Fligth<Predicate>. This functions will be named after a "flight
predicate", as for example "routeIs".

Idea is to be able to write (with a proper static import):
   flightSource.getFlights(routeIs("from","to"))

Created the routeIs function inside a RouteIs container class.

Adapt the tests to the new syntax (note the static import).

Remove RouteFligthFilter.