Refactor to be able to test Flight instantation from List<String>
Use template method pattern.

Abstract class StreamOfListOfStringFlightDeserializer constains the
logic to create a Flight instance from the List<String>. If the
list has less than 3 items (the minimum needed to create a flight,
origin, destination and code), then returns null. If list has
more items, then they are ignored.

This logic is used in a call to Stream.map function to build a stream
of flights, which is then filtered to remove nulls.

Added tests to verify this logic in different scenarions

CsvFileFlightSource implements the abstract method readFlightData
by calling the CsvFiles utility.