FlightSearchResult: price is optional
Allow the possibility of havig a search result in which the price
has not been set (e.g, because the price could not be retrieved
from the pricing source).

Adjust the equals function and add tests for it as the logic has a
bit more complexity now.

Note that we do not want to use the Optional<BigDecimal>.equals
function because it will delegate in BigDecimal.equals, and this
function will not consider equal two BidDecimals with same value
but different scale. Therefore we need to use BigDecimal.compareTo.