Refactor CsvFilePricingSource
Create abstract class StreamOfListOfStringPriceDeserializer, with
abstract method to get a Stream<List<String>>

Use template method pattern: CsvFilePricingSource provides an
implementatin of the abstract method that actually reads from
a CSV file using the CsvFiles library.

Added error handling: the optional returned will be empty if no price
was found for the given flight code, or if the entry has not enough
fields to read the price from, or if the price value cannot be parsed to
a double value.